pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gdrivectl"
version = "0.1.0"
description = "CLI tool to manage Google Docs permissions in bulk — grant, revoke, audit"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Kushagra Sharma", email = "kushagrasharma.dev@gmail.com" },
]
keywords = ["google-drive", "permissions", "cli", "google-docs"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Utilities",
]
dependencies = [
    "google-api-python-client>=2.0",
    "google-auth>=2.0",
    "google-auth-oauthlib>=1.0",
    "click>=8.0",
    "questionary>=2.0",
    "rich>=13.0",
]

[project.urls]
Homepage = "https://github.com/TheDarkArtist/gdrivectl"
Repository = "https://github.com/TheDarkArtist/gdrivectl"
Issues = "https://github.com/TheDarkArtist/gdrivectl/issues"

[project.scripts]
gdrivectl = "gdrivectl.cli:cli"
